Business refactor beginning
